requirements.txt
feedparser~=6.0.11
pydantic~=2.11.7
motor~=3.7.1
langchain-openai~=0.3.28
langchain~=0.3.27
dotenv~=0.9.9
python-dotenv~=1.1.1
azure-core~=1.35.0
Quart~=0.20.0
aiohttp~=3.12.15
pydantic-settings~=2.10.1
fastapi~=0.116.1
config~=0.5.1
uvicorn~=0.35.0
openai~=1.98.0